requirements.txt
Django==2.1.1
django-crispy-forms==1.7.2
numpy==1.15.1
pandas==0.23.4
Pillow==5.2.0
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
manage